+++ src/StatsdExporterClientInterface.php
@@ -5,7 +5,7 @@ namespace TutuRu\Metrics;
 
 interface StatsdExporterClientInterface
 {
-    public function count(string $key, int $value, array $tags = []): StatsdExporterClientInterface;
+    public function count(string $key, float $value, array $tags = []): StatsdExporterClientInterface;
 
     public function increment(string $key, array $tags = []): StatsdExporterClientInterface;
 
@@ -13,9 +13,9 @@ interface StatsdExporterClientInterface
 
     public function timing(string $key, float $seconds, array $tags = []): StatsdExporterClientInterface;
 
-    public function gauge(string $key, int $value, array $tags = []): StatsdExporterClientInterface;
+    public function gauge(string $key, float $value, array $tags = []): StatsdExporterClientInterface;
 
-    public function summary(string $key, int $value, array $tags = []): StatsdExporterClientInterface;
+    public function summary(string $key, float $value, array $tags = []): StatsdExporterClientInterface;
 
     public function save(): void;
 }
